Add a required permission annotation with a permission string value to each endpoint
