+++ back-end/auth-service/src/main/java/com/ecommerce/auth_service/controller/UserController.java
@@ -5,6 +5,7 @@ import com.ecommerce.auth_service.model.request.RegistrationUserRequest;
 import com.ecommerce.auth_service.model.request.UpdateUserRequest;
 import com.ecommerce.auth_service.model.response.BaseResponse;
 import com.ecommerce.auth_service.model.response.PaginationInfo;
+import com.ecommerce.auth_service.security.RequiredPermission;
 import com.ecommerce.auth_service.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -20,19 +21,14 @@ import javax.validation.Valid;
 public class UserController {
     private final UserService userService;
 
-    @PostMapping(value = "/register", consumes = "application/json", produces = "application/json")
-    public ResponseEntity<BaseResponse> registerUser(@Valid @RequestBody RegistrationUserRequest request) {
-        userService.register(request);
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .body(new BaseResponse(HttpStatus.CREATED.toString(), "User registered successfully", null, null));
-    }
-
+    @RequiredPermission("user:read")
     @GetMapping("/{id}")
     public ResponseEntity<BaseResponse> getUserById(@PathVariable("id") Long id) {
         User user = userService.findById(id);
         return ResponseEntity.ok(new BaseResponse(HttpStatus.OK.toString(), "User detail", user, null));
     }
 
+    @RequiredPermission("user:read")
     @GetMapping()
     public ResponseEntity<BaseResponse> findUsers(@RequestParam(defaultValue = "1") int page,
                                                   @RequestParam(defaultValue = "10") int size) {
@@ -48,18 +44,21 @@ public class UserController {
         ));
     }
 
+    @RequiredPermission("user:delete")
     @DeleteMapping("/{id}")
     public ResponseEntity<BaseResponse> inactivateUser(@PathVariable("id") Long id) {
         userService.deactivate(id);
         return ResponseEntity.ok(new BaseResponse(HttpStatus.OK.toString(), "User inactivate successfully", null, null));
     }
 
+    @RequiredPermission("user:activate")
     @PostMapping("/{id}")
     public ResponseEntity<BaseResponse> requestReactivation(@PathVariable("id") Long id) {
         userService.requestUserReactivation(id);
         return ResponseEntity.ok(new BaseResponse(HttpStatus.OK.toString(), "request user reactive successfully", null, null));
     }
 
+    @RequiredPermission("user:write")
     @PatchMapping("/{id}")
     public ResponseEntity<BaseResponse> updateUser(@PathVariable("id") Long id, @RequestBody UpdateUserRequest request) {
         userService.updateUser(id, request);
